requirements.txt
translate==3.6.1
gTTS==2.5.3
librosa==0.10.2.post1
matplotlib==3.9.2
numpyc
scipy==1.14.1
sounddevice==0.5.0
SpeechRecognition==3.10.4
streamlit==1.40.0
tensorflow==2.18.0
PyAudio==0.2.14